feat: Modal의 네이밍을 ModalMain으로 수정하고, props 방식에서 합성 컴포넌트 방식으로 구조 변경
