+++ components/src/App.tsx
@@ -1,7 +1,6 @@
 import React, { useReducer } from 'react';
 
-import { Modal } from 'maru-nice-modal';
-import SelectBank from './components/SelectBank';
+import { Modal } from './lib';
 
 function App() {
   const [isOpen, toggleIsOpen] = useReducer(prev => !prev, false);
@@ -9,16 +8,7 @@ function App() {
   return (
     <>
       <button onClick={toggleIsOpen}>모달열기</button>
-      <Modal
-        isOpen={isOpen}
-        content={<SelectBank />}
-        onClose={() => toggleIsOpen()}
-        title="카드사 선택"
-        position="center"
-        closeButton={<button style={{ width: '100%' }}>닫기 버튼</button>}
-        confirmButton={<button style={{ width: '100%' }}>확인 버튼</button>}
-        buttonPosition="column"
-      />
+      <Modal isOpen={isOpen}></Modal>
     </>
   );
 }
